feat(auth): implement JWT authentication with login and refresh token functionality
- Added @nestjs/jwt package for JWT handling.
- Created login and refresh token DTOs for authentication.
- Implemented AuthService methods for login and token refresh.
- Updated AuthController to handle login and refresh token routes.
- Introduced JWT configuration for secret and expiration settings.
- Added role-based access control with RolesGuard and AuthAndPolicyGuard.
- Refactored UserService to include role checks for user actions.
- Removed unused CreateAuthDto and UpdateAuthDto.
- Enhanced user entity to include roles.
- Added decorators for route policy management.
- Updated response DTOs to include necessary fields.

+++ back-end/src/user/user.service.ts
@@ -1,5 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { SetRoutePolicy } from 'src/auth/decorators/set-route-policy.decorator';
+import { PayloadDto } from 'src/auth/dto/payload.dto';
+import { Roles } from 'src/auth/enums/roles.enums';
 import { HashingProtocol } from 'src/auth/hashing/hashing-protocol';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -22,35 +25,55 @@ export class UserService {
     return this.userMapper.toResponse(user);
   }
 
+  @SetRoutePolicy(Roles.Admin)
   async findAll() {
     const allUsers = await this.userRepository.find();
     return allUsers.map((user) => this.userMapper.toResponse(user));
   }
 
-  async findOne(id: string) {
+  @SetRoutePolicy(Roles.Admin, Roles.User)
+  async findOne(id: string, payload: PayloadDto) {
     const user = await this.userRepository.findOneBy({ id });
     if (!user) {
       throw new Error('User not found');
+    }
+    if (user.id !== payload.sub) {
+      throw new Error('Unauthorized');
     }
     return this.userMapper.toResponse(user);
   }
 
-  // TODO: verificar se realmente eu quero que o usuário seja atualizado por completo sem verificar o que vem no DTO
-  async update(id: string, updateUserDto: UpdateUserDto) {
+  @SetRoutePolicy(Roles.Admin, Roles.User)
+  async update(
+    id: string,
+    updateUserDto: UpdateUserDto,
+    tokenPayload: PayloadDto,
+  ) {
     const user = await this.userRepository.findOneBy({ id });
+
     if (!user) {
       throw new Error('User not found');
     }
+
+    if (user.id !== tokenPayload.sub) {
+      throw new Error('Unauthorized');
+    }
+
     this.userRepository.merge(user, updateUserDto);
     await this.userRepository.save(user);
     return this.userMapper.toResponse(user);
   }
 
-  async remove(id: string) {
+  @SetRoutePolicy(Roles.Admin, Roles.User)
+  async remove(id: string, payload: PayloadDto) {
     const user = await this.userRepository.findOneBy({ id });
     if (!user) {
       throw new Error('User not found');
     }
+    if (user.id !== payload.sub) {
+      throw new Error('Unauthorized');
+    }
+
     return await this.userRepository.remove(user);
   }
 }
